EW-949 code cleanup
