+++ cypress/support/custom_commands/login.js
@@ -1,4 +1,8 @@
-const { loginWithoutSchoolApi, loginViaSchoolApi } = require("./login.helper");
+import {
+	loginWithoutSchoolApi,
+	loginViaSchoolApi,
+	getUserCredentials,
+} from "./login.helper";
 
 const initials = '[data-testid="initials"]';
 const languageSelection = '[id="selected-language"]';
@@ -22,4 +26,30 @@ Cypress.Commands.add("login", (username, environment) => {
 	});
 	cy.visit("/dashboard");
 	cy.get(titleOnDashboardPage).should("exist");
+});
+
+Cypress.Commands.add("apiLogin", (user, environment) => {
+	const env = Cypress.env();
+	const environmentUpperCased = environment.toUpperCase();
+
+	Cypress.config("baseUrl", env[environmentUpperCased]);
+
+	cy.session([user, environment], () => {
+		const [username, password] = getUserCredentials(user);
+
+		cy.request({
+			method: "POST",
+			url: getPageUrl(environment, "/api/v3/authentication/local"),
+			body: {
+				username: env[username],
+				password: env[password],
+			},
+		}).then(({ body }) => {
+			const domain = new URL(Cypress.config("baseUrl")).hostname;
+
+			cy.setCookie("jwt", body.accessToken, { domain });
+		});
+	});
+
+	cy.visitPage(environment, "/dashboard");
 });
